fix(agent): break circular import; fix LLM session test sink
CI surfaced two failures on the previous test commit:

1. Circular module-load: agent-runtime.ts instantiates
   `new LlmAgentSdkProvider()` at module top level, but llm-agent.ts
   imports SYSTEM_PROMPT / AgentQuerySession / AgentSdkProvider from
   agent-runtime.ts — so when the test file imports llm-agent.ts the
   class binding is undefined by the time the runtime evaluates its
   provider map (TDZ-style: "LlmAgentSdkProvider is not a constructor").

   Fix: extract SYSTEM_PROMPT and the two interfaces into a new
   `sdk-provider.ts`. agent-runtime.ts re-exports them so existing
   import sites keep working; llm-agent.ts now imports from
   sdk-provider.js, breaking the cycle.

2. The "refuses to resume a thread that belongs to another user" test
   asserted on transport.streamMessage being called, but
   LlmAgentSession.send() returns the messages and emits via the
   onMessage callback — transport.streamMessage is the next layer up
   (AgentRuntime.sendMessageStreaming). Switched the assertion to read
   the returned message array, which is the actual session contract.

Also updates an inherited test:
- runtime/tests/providers/openai-provider.test.ts expected 3 image
  models; the OpenAI provider now lists 4 (gpt-image-2 was added in
  4b5e1bb without updating this assertion). Bumping the expectation
  unblocks Packages Test on this branch.

All 70 websocket tests + 10 AgentStore tests + 3 useModelsByProvider
tests now pass locally.
