+++ packages/runtime/tests/providers/openai-provider.test.ts
@@ -198,7 +198,7 @@ describe("OpenAIProvider", () => {
 
     await expect(provider.getAvailableASRModels()).resolves.toHaveLength(1);
     await expect(provider.getAvailableTTSModels()).resolves.toHaveLength(2);
-    await expect(provider.getAvailableImageModels()).resolves.toHaveLength(3);
+    await expect(provider.getAvailableImageModels()).resolves.toHaveLength(4);
     await expect(provider.getAvailableVideoModels()).resolves.toHaveLength(2);
     await expect(provider.getAvailableEmbeddingModels()).resolves.toHaveLength(
       3
